requirements.txt
# Core dependencies for Polymarket Arbitrage Bot
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
requests==2.31.0
requests-cache>=1.1.0
PyYAML>=6.0
rich>=13.0.0
python-dateutil>=2.8.2
pytz>=2023.3
werkzeug>=3.0.0

# Crypto APIs and Data Sources
cachetools==5.3.1
websocket-client>=1.6.4
gql>=3.4.1
requests-toolbelt>=1.0.0

# Analytics and Data Processing
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3

# Async Operations
aiohttp>=3.9.0

# System Monitoring
psutil>=5.9.0
plyer>=2.1.0

# Email and Communication
secure-smtplib>=0.1.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.4.1

# Documentation
sphinx==7.1.2

# Production
gunicorn==22.0.0
alembic==1.12.0

# Database
SQLAlchemy==2.0.21

# Monitoring and Observability
prometheus-client>=0.18.0
sentry-sdk>=1.38.0

# Security
PyJWT>=2.8.0
cryptography>=41.0.0

# Caching and Performance
redis>=5.0.0
hiredis>=2.2.0
